client/requirements.txt
azure-cognitiveservices-speech
sounddevice
python-dotenv
requests
